Conflicts:
	notib-core-api/src/main/java/es/caib/notib/core/api/dto/PagadorCieDto.java
	notib-core/src/main/java/es/caib/notib/core/helper/ConversioTipusHelper.java

+++ notib-war/src/main/java/es/caib/notib/war/controller/GrupController.java
@@ -112,7 +112,7 @@ public class GrupController extends BaseUserController{
 			return getModalControllerReturnValueSuccess(
 					request,
 					"redirect:grupAdminList",
-					"procediment.controller.modificat.ok");
+					"grup.controller.modificat.ok");
 		//if it is new	
 		} else {
 			grupService.create(
@@ -121,7 +121,7 @@ public class GrupController extends BaseUserController{
 			return getModalControllerReturnValueSuccess(
 					request,
 					"redirect:grupAdminList",
-					"procediment.controller.creat.ok");
+					"grup.controller.creat.ok");
 		}
 	}
 	
@@ -155,8 +155,15 @@ public class GrupController extends BaseUserController{
 	public String delete(
 			HttpServletRequest request,
 			@PathVariable Long grupId) {
-		//EntitatDto entitatActual = getEntitatActualComprovantPermisos(request);
+		EntitatDto entitatActual = getEntitatActualComprovantPermisos(request);
 		
+		// Comprova que el grup no s'utilitzi
+		if (grupService.existProcedimentGrupByGrupId(entitatActual.getId(), grupId)) {
+			return getAjaxControllerReturnValueError(
+					request, 
+					"redirect:../../grup", 
+					"grup.controller.esborrat.ko.enus");
+		};
 		grupService.delete(grupId);
 		
 		return getAjaxControllerReturnValueSuccess(
